+++ src/main/java/com/marginallyclever/robotOverlord/arm5/tool/EvilMinionTool.java
@@ -39,5 +39,5 @@ public abstract class EvilMinionTool extends PhysicalObject {
 	
 	public void updateGUI() {}
 	
-	public void update(float dt) {}
+	public void update(double dt) {}
 }
