+++ src/main/java/com/marginallyclever/robotOverlord/spidee/SpideeLocation.java
@@ -2,7 +2,7 @@ package com.marginallyclever.robotOverlord.spidee;
 
 import java.io.Serializable;
 
-import javax.vecmath.Vector3f;
+import javax.vecmath.Vector3d;
 
 public class SpideeLocation implements Serializable {
 	  /**
@@ -10,9 +10,9 @@ public class SpideeLocation implements Serializable {
 	 */
 	private static final long serialVersionUID = -1812312249185841160L;
 	
-	Vector3f up = new Vector3f();
-	  Vector3f left = new Vector3f();
-	  Vector3f forward = new Vector3f();
-	  Vector3f pos = new Vector3f();
-	  Vector3f relative = new Vector3f();
+	Vector3d up = new Vector3d();
+	Vector3d left = new Vector3d();
+	Vector3d forward = new Vector3d();
+	Vector3d pos = new Vector3d();
+	Vector3d relative = new Vector3d();
 }
